+++ src/chain/listener.py
@@ -12,7 +12,7 @@ from __future__ import annotations
 
 import asyncio
 from asyncio import create_task, sleep
-from typing import Optional, cast
+from typing import cast
 
 from eth_typing import BlockNumber
 from reretry import retry  # type: ignore
@@ -23,14 +23,12 @@ from chain.reader import Reader
 from chain.registry import Registry
 from chain.rpc import RPC
 from orchestration.guardian import Guardian
+from shared.config import ConfigSnapshotSync
 from shared.message import GuardianError, SubscriptionCreatedMessage
 from shared.service import AsyncTask
 from utils import log
 
-SNAPSHOT_SYNC_BATCH_SIZE = 200
-SNAPSHOT_SYNC_BATCH_SLEEP_S = 1.0
 SUBSCRIPTION_SYNC_BATCH_SIZE = 20
-SNAPSHOT_SYNC_STARTING_SUB_ID = 0
 
 
 def get_batches(start: int, end: int, batch_size: int) -> list[tuple[int, int]]:
@@ -72,6 +70,8 @@ class ChainListener(AsyncTask):
         _trail_head_blocks (int): How many blocks to trail head by
         _snapshot_sync_sleep (int): Snapshot sync sleep time between each batch
         _snapshot_sync_batch_size (int): Snapshot sync batch size to sync in parallel
+        _snapshot_sync_starting_sub_id (int): Snapshot sync starting subscription ID
+        _syncing_period (float): How long to sleep between each iteration
     """
 
     def __init__(
@@ -83,9 +83,7 @@ class ChainListener(AsyncTask):
         guardian: Guardian,
         processor: ChainProcessor,
         trail_head_blocks: int,
-        snapshot_sync_sleep: Optional[int],
-        snapshot_sync_batch_size: Optional[int],
-        snapshot_sync_starting_sub_id: Optional[int],
+        snapshot_sync: ConfigSnapshotSync,
     ) -> None:
         """Initializes new ChainListener
 
@@ -97,8 +95,7 @@ class ChainListener(AsyncTask):
             guardian (Guardian): Guardian instance
             processor (ChainProcessor): ChainProcessor instance
             trail_head_blocks (int): How many blocks to trail head by
-            snapshot_sync_sleep (int): Snapshot sync sleep time between each batch
-            snapshot_sync_batch_size (int): Snapshot sync batch size to sync in parallel
+            snapshot_sync (ConfigSnapshotSync): Snapshot sync configuration
         """
 
         # Initialize inherited AsyncTask
@@ -111,22 +108,10 @@ class ChainListener(AsyncTask):
         self._guardian = guardian
         self._processor = processor
         self._trail_head_blocks = trail_head_blocks
-        self._snapshot_sync_sleep = (
-            SNAPSHOT_SYNC_BATCH_SLEEP_S
-            if snapshot_sync_sleep is None
-            else snapshot_sync_sleep
-        )
-        self._snapshot_sync_batch_size = (
-            SNAPSHOT_SYNC_BATCH_SIZE
-            if snapshot_sync_batch_size is None
-            else snapshot_sync_batch_size
-        )
-        self._snapshot_sync_starting_sub_id = (
-            SNAPSHOT_SYNC_STARTING_SUB_ID
-            if snapshot_sync_starting_sub_id is None
-            else snapshot_sync_starting_sub_id
-        )
-
+        self._snapshot_sync_sleep = snapshot_sync.sleep
+        self._snapshot_sync_batch_size = snapshot_sync.batch_size
+        self._snapshot_sync_starting_sub_id = snapshot_sync.starting_sub_id
+        self._syncing_period = snapshot_sync.sync_period
         log.info("Initialized ChainListener")
 
     async def _sync_batch_subscriptions_creation(
@@ -188,23 +173,24 @@ class ChainListener(AsyncTask):
                     if subscription.id == sub_id:
                         subscription.set_response_count(interval, response_count)
                         # Create new subscription created message
-                        msg = SubscriptionCreatedMessage(subscription)
 
-                        # Run message through guardian
-                        filtered = self._guardian.process_message(msg)
+            for subscription in subscriptions:
+                msg = SubscriptionCreatedMessage(subscription)
 
-                        if isinstance(filtered, GuardianError):
-                            # If filtered out by guardian, message is irrelevant
-                            log.info(
-                                "Ignored subscription creation",
-                                id=sub_id,
-                                err=filtered.error,
-                            )
-                        else:
-                            # Pass filtered message to ChainProcessor
-                            create_task(self._processor.track(msg))
-                            log.info("Relayed subscription creation", id=sub_id)
-                        break
+                # Run message through guardian
+                filtered = self._guardian.process_message(msg)
+
+                if isinstance(filtered, GuardianError):
+                    # If filtered out by guardian, message is irrelevant
+                    log.info(
+                        "Ignored subscription creation",
+                        id=sub_id,
+                        err=filtered.error,
+                    )
+                else:
+                    # Pass filtered message to ChainProcessor
+                    create_task(self._processor.track(msg))
+                    log.info("Relayed subscription creation", id=subscription.id)
             break
         return
 
@@ -334,7 +320,7 @@ class ChainListener(AsyncTask):
                 log.info(f"head sub id is: {head_sub_id}")
                 num_subs_to_sync = min(
                     head_sub_id - self._last_subscription_id,
-                    SUBSCRIPTION_SYNC_BATCH_SIZE,
+                    self._snapshot_sync_batch_size,
                 )
 
                 # Collect all Coordinator emitted event logs in range
@@ -362,12 +348,14 @@ class ChainListener(AsyncTask):
             else:
                 # Else, if already synced to head, sleep
                 log.debug(
-                    "No new blocks, sleeping for 500ms",
+                    "No new blocks, sleeping for: ",
+                    self._syncing_period,
+                    "seconds",
                     head=head_block,
                     synced=self._last_block,
                     behind=self._trail_head_blocks,
                 )
-                await sleep(0.5)
+                await sleep(self._syncing_period)
 
     async def cleanup(self: ChainListener) -> None:
         """Stateless task, no cleanup necessary"""