+++ src/utils/logging.py
@@ -1,19 +1,18 @@
 import logging
-import logging.handlers
-from typing import Literal, Optional
+from logging.handlers import RotatingFileHandler
+from typing import Literal
 
 import pyfiglet  # type: ignore
 import structlog
 from rich import print
-from structlog.typing import Processor
 
-from utils.config import ConfigLog
+from shared.config import ConfigLog
 
 # Re-export logger
 log = structlog.get_logger()
 
 # Structlog shared processors
-SHARED_PROCESSORS: list[Processor] = [
+SHARED_PROCESSORS: list[structlog.typing.Processor] = [
     structlog.contextvars.merge_contextvars,  # Merge in global context
     structlog.stdlib.add_log_level,  # Add log level
     structlog.stdlib.add_logger_name,  # Add logging function
@@ -25,23 +24,13 @@ SHARED_PROCESSORS: list[Processor] = [
 # Font for ASCII art, taken from http://www.figlet.org/examples.html
 PIGLET_FONT = "o8"
 
-DEFAULT_LOG_PATH = "infernet_node.log"
-DEFAULT_MAX_FILE_SIZE = 2**30  # Default to 1GB log file size
-DEFAULT_BACKUP_COUNT = 2  # Default to 2 log files to keep
 
-
-def setup_logging(
-    config: Optional[ConfigLog],
-) -> None:
+def setup_logging(config: ConfigLog) -> None:
     """Setup logging configuration
 
     Args:
-        config (Optional[ConfigLog]): Logging configuration options
+        config (ConfigLog): Logging configuration options
     """
-
-    path = config.get("path") if config else None
-    max_file_size = config.get("max_file_size") if config else None
-    backup_count = config.get("backup_count") if config else None
 
     # Configure structlog
     # Largely standard config: https://www.structlog.org/en/stable/configuration.html
@@ -60,10 +49,10 @@ def setup_logging(
     console_handler = logging.StreamHandler()  # Stream to sys.stderr
 
     # Use RotatingFileHandler to limit log file size
-    file_handler = logging.handlers.RotatingFileHandler(
-        DEFAULT_LOG_PATH if path is None else path,
-        maxBytes=DEFAULT_MAX_FILE_SIZE if max_file_size is None else max_file_size,
-        backupCount=DEFAULT_BACKUP_COUNT if backup_count is None else backup_count,
+    file_handler = RotatingFileHandler(
+        config.path,
+        maxBytes=config.max_file_size,
+        backupCount=config.backup_count,
     )
 
     # Setup log formatting