+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 - ##### The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - ##### This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2024-XX-XX
+
+### Added
+
+- New `sync_period` field as `snapshot_sync` configuration parameter, to optionally set the period of polling the chain for new blocks/subscriptions.
+
+### Changed
+
+- Almost all container configurations (except `id` and `image`) are now reasonably defaulted (e.g. port auto-assignment), making it easier to configure a service and enabling the use of service recipes. Container configurations are validated at startup and errors are reported early.
+- Added validation for entire `config.json` using pydantic, for sane defaulting and more meaningful error messages.
+
+### Fixed
+
+- Bug with batch syncing new subscriptions
+
 ## [1.2.0] - 2024-08-15
 
 ### Added